+++ repomaestro/platform/github.py
@@ -24,10 +24,12 @@ def get_repos_data(github_token: str, github_ids: list) -> dict:
     for repo in g.get_user().get_repos():
         if github_ids == [] or repo.owner.login in github_ids:
             logger.info(f"- {repo.name}")
+            # Complete list of PyGithub Repository: https://pygithub.readthedocs.io/en/stable/github_objects/Repository.html
             data[repo.name] = {
                 "homepage": repo.homepage,
                 "keywords": repo.topics,
                 "git_url": repo.git_url,
+                "ssh_url": repo.ssh_url
             }
         else:
             logger.info(f"- {repo.name} (skipped)")