+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add ssh_url to repository properties map
+
 ## 1.0.0 - 2025-01-06
 ### Added
 - Add --exclude-keywords flag to gen-file command