+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix ssh_url due to empty value from PyGithub.Repository
+
 ## 1.1.0 - 2025-04-13
 ### Added
 - Add ssh_url to repository properties map